In the beginning god created the heavens and the earth
ok so i made most of these files a while ago, but I never got a very good handle on vulkan.
This is a functional(?) hello triangle in vulkan.
It works on my machine, anyways.

There's a lot of tidying to do, and optimizing, and making the code valid other than by coincidence.
This uses my makefile and a lot of the filesfrom my "helloworld" project,
which is why I've uncreatively called this "helloworld_sdl".
I have already played with using helloworld to deal with sdl, but I wanted to
see if I could strap vulkan on it's back as well, and it's turning out OK.