Lots of changes
I'm in the process of isolating things to make application development less messy.
The function write_command_buffers in triangle.c is called elsewhere. it shouldn't be, but it must be right now.
I'm going to work to change that by using the newer makefile setup with modules, and making a similar system for SDL modules!

More to come soon :)